+++ src/Repository/ElasticsearchRepository.php
@@ -3,12 +3,14 @@ declare(strict_types=1);
 
 namespace Kununu\Elasticsearch\Repository;
 
+use Elasticsearch\Client;
 use Exception;
-use Kununu\Elasticsearch\Adapter\AdapterFactoryInterface;
 use Kununu\Elasticsearch\Exception\RepositoryException;
 use Kununu\Elasticsearch\Query\Query;
 use Kununu\Elasticsearch\Query\QueryInterface;
 use Kununu\Elasticsearch\Result\AggregationResultSet;
+use Kununu\Elasticsearch\Result\AggregationResultSetInterface;
+use Kununu\Elasticsearch\Result\ResultIterator;
 use Kununu\Elasticsearch\Result\ResultIteratorInterface;
 use Kununu\Elasticsearch\Util\LoggerAwareTrait;
 use Psr\Log\LoggerAwareInterface;
@@ -25,25 +27,25 @@ class ElasticsearchRepository implements ElasticsearchRepositoryInterface, Logge
     protected const EXCEPTION_PREFIX = 'Elasticsearch exception: ';
 
     /**
-     * @var \Kununu\Elasticsearch\Adapter\AdapterInterface
+     * @var \Elasticsearch\Client
      */
     protected $client;
 
     /**
-     * @var array
+     * @var \Kununu\Elasticsearch\Repository\RepositoryConfiguration
      */
-    protected $connectionConfig;
+    protected $config;
 
     /**
-     * AbstractElasticsearchManager constructor.
+     * ElasticsearchRepository constructor.
      *
-     * @param \Kununu\Elasticsearch\Adapter\AdapterFactoryInterface $adapterFactory
-     * @param array                                                 $connectionConfig
+     * @param \Elasticsearch\Client $client
+     * @param array                 $config
      */
-    public function __construct(AdapterFactoryInterface $adapterFactory, array $connectionConfig)
+    public function __construct(Client $client, array $config)
     {
-        $this->client = $adapterFactory->build($connectionConfig['adapter_class'] ?? '', $connectionConfig);
-        $this->connectionConfig = $connectionConfig;
+        $this->client = $client;
+        $this->config = new RepositoryConfiguration($config);
     }
 
     /**
@@ -59,12 +61,75 @@ class ElasticsearchRepository implements ElasticsearchRepositoryInterface, Logge
     }
 
     /**
+     * @param string $operationType
+     *
+     * @return array
+     */
+    protected function buildRequestBase(string $operationType): array
+    {
+        return [
+            'index' => $this->config->getIndex($operationType),
+            'type' => $this->config->getType(),
+        ];
+    }
+
+    /**
+     * @param \Kununu\Elasticsearch\Query\QueryInterface $query
+     * @param string                                     $operationType
+     *
+     * @return array
+     */
+    protected function buildRawQuery(QueryInterface $query, string $operationType): array
+    {
+        return array_merge(
+            $this->buildRequestBase($operationType),
+            ['body' => $query->toArray()]
+        );
+    }
+
+    /**
+     * @param array $rawResult
+     *
+     * @return \Kununu\Elasticsearch\Result\ResultIteratorInterface
+     */
+    protected function parseRawSearchResponse(array $rawResult): ResultIteratorInterface
+    {
+        return ResultIterator::create($rawResult['hits']['hits'] ?? [])
+            ->setTotal($rawResult['hits']['total'] ?? 0)
+            ->setScrollId($rawResult['_scroll_id'] ?? null);
+    }
+
+    /**
+     * @param array $updateScript
+     *
+     * @return array
+     */
+    protected function sanitizeUpdateScript(array $updateScript): array
+    {
+        if (!isset($updateScript['script']) && count($updateScript) > 1) {
+            $sanitizedUpdateScript = [
+                'script' => [
+                    'lang' => $updateScript['lang'] ?? null,
+                    'source' => $updateScript['source'] ?? [],
+                    'params' => $updateScript['params'] ?? [],
+                ],
+            ];
+        } else {
+            $sanitizedUpdateScript = $updateScript;
+        }
+
+        return $sanitizedUpdateScript;
+    }
+
+    /**
      * @inheritdoc
      */
     public function save(string $id, array $document): void
     {
         try {
-            $this->client->index($id, $document);
+            $this->client->index(
+                array_merge($this->buildRequestBase(OperationType::WRITE), ['id' => $id, 'body' => $document])
+            );
         } catch (Exception $e) {
             $this->logErrorAndThrowException($e);
         }
@@ -76,7 +141,9 @@ class ElasticsearchRepository implements ElasticsearchRepositoryInterface, Logge
     public function delete(string $id): void
     {
         try {
-            $this->client->delete($id);
+            $this->client->delete(
+                array_merge($this->buildRequestBase(OperationType::WRITE), ['id' => $id])
+            );
         } catch (Exception $e) {
             $this->logErrorAndThrowException($e);
         }
@@ -88,7 +155,7 @@ class ElasticsearchRepository implements ElasticsearchRepositoryInterface, Logge
     public function deleteIndex(string $indexName): void
     {
         try {
-            $this->client->deleteIndex($indexName);
+            $this->client->indices()->delete(['index' => $indexName]);
         } catch (Exception $e) {
             $this->logErrorAndThrowException($e);
         }
@@ -100,7 +167,9 @@ class ElasticsearchRepository implements ElasticsearchRepositoryInterface, Logge
     public function findByQuery(QueryInterface $query): ResultIteratorInterface
     {
         try {
-            return $this->client->search($query);
+            return $this->parseRawSearchResponse(
+                $this->client->search($this->buildRawQuery($query, OperationType::READ))
+            );
         } catch (Exception $e) {
             $this->logErrorAndThrowException($e);
         }
@@ -114,7 +183,12 @@ class ElasticsearchRepository implements ElasticsearchRepositoryInterface, Logge
     public function findScrollableByQuery(QueryInterface $query): ResultIteratorInterface
     {
         try {
-            return $this->client->search($query, true);
+            $rawQuery = $this->buildRawQuery($query, OperationType::READ);
+            $rawQuery['scroll'] = $this->config->getScrollContextKeepalive();
+
+            return $this->parseRawSearchResponse(
+                $this->client->search($rawQuery)
+            );
         } catch (Exception $e) {
             $this->logErrorAndThrowException($e);
         }
@@ -126,7 +200,14 @@ class ElasticsearchRepository implements ElasticsearchRepositoryInterface, Logge
     public function findByScrollId(string $scrollId): ResultIteratorInterface
     {
         try {
-            return $this->client->scroll($scrollId);
+            return $this->parseRawSearchResponse(
+                $this->client->scroll(
+                    [
+                        'scroll_id' => $scrollId,
+                        'scroll' => $this->config->getScrollContextKeepalive(),
+                    ]
+                )
+            );
         } catch (Exception $e) {
             $this->logErrorAndThrowException($e);
         }
@@ -146,7 +227,7 @@ class ElasticsearchRepository implements ElasticsearchRepositoryInterface, Logge
     public function countByQuery(QueryInterface $query): int
     {
         try {
-            return $this->client->count($query);
+            return $this->client->count($this->buildRawQuery($query, OperationType::READ))['count'];
         } catch (Exception $e) {
             $this->logErrorAndThrowException($e);
         }
@@ -155,10 +236,15 @@ class ElasticsearchRepository implements ElasticsearchRepositoryInterface, Logge
     /**
      * @inheritdoc
      */
-    public function aggregateByQuery(QueryInterface $query): AggregationResultSet
+    public function aggregateByQuery(QueryInterface $query): AggregationResultSetInterface
     {
         try {
-            return $this->client->aggregate($query);
+            $result = $this->client->search(
+                $this->buildRawQuery($query, OperationType::READ)
+            );
+
+            return AggregationResultSet::create($result['aggregations'] ?? [])
+                ->setDocuments($this->parseRawSearchResponse($result));
         } catch (Exception $e) {
             $this->logErrorAndThrowException($e);
         }
@@ -170,7 +256,10 @@ class ElasticsearchRepository implements ElasticsearchRepositoryInterface, Logge
     public function updateByQuery(QueryInterface $query, array $updateScript): array
     {
         try {
-            return $this->client->update($query, $updateScript);
+            $rawQuery = $this->buildRawQuery($query, OperationType::WRITE);
+            $rawQuery['body']['script'] = $this->sanitizeUpdateScript($updateScript)['script'];
+
+            return $this->client->updateByQuery($rawQuery);
         } catch (Exception $e) {
             $this->logErrorAndThrowException($e);
         }