+++ src/Repository/OperationType.php
@@ -1,0 +1,15 @@
+<?php
+declare(strict_types=1);
+
+namespace Kununu\Elasticsearch\Repository;
+
+/**
+ * Class OperationType
+ *
+ * @package Kununu\Elasticsearch\Repository
+ */
+final class OperationType
+{
+    public const READ = 'read';
+    public const WRITE = 'write';
+}